refactor(eve): share agent-config AST plumbing and the model-summary grammar
Both source editors (apply-model-name and agent-config-string-path)
carried their own copies of the Rolldown parse, defineAgent lookup,
expression unwrapping, key matching, and quote escaping — ~150 lines
drifting independently. Extract them into agent-config-ast.ts; each
editor keeps only its lookup policy (the model rewriter stays
deliberately permissive about spreads, the path editor stays strict).
checkAgentConfigSource moves to the shared module and is re-exported
from agent-config-string-path so existing imports hold.

The slug@level ↯ summary grammar was likewise written twice — the dev
TUI status line and the /model Change-model hint. formatModelSummary
in shared/model-summary.ts is now the single implementation, so the
two surfaces cannot drift.
